+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 
 ### Changed
-
+* [ALFREDAPI-395](https://xenitsupport.jira/com/browse/ALFREDAPI-395): Change loglevels of PermissionSerivce#hasPermission
 
 ### Fixed
 * [ALFREDAPI-428](https://xenitsupport.jira.com/browse/ALFREDAPI-428): Replace PersonService#getPerson(String userName) with getPersonOrNull(String userName) to avoid creation of new users when getPerson is called with a nonexistent userName.