+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Alfred API - Changelog
 
-## 2.5.3 UNRELEASED
+## 2.6.0 (UNRELEASED)
 
 ### Added
-
+* [ALFREDAPI-298](https://xenitsupport.jira.com/browse/ALFREDAPI-438): Add CRUD api for comments
 
 ### Changed
 * [ALFREDAPI-395](https://xenitsupport.jira/com/browse/ALFREDAPI-395): Change loglevels of PermissionSerivce#hasPermission
@@ -11,11 +11,16 @@
 ### Fixed
 * [ALFREDAPI-428](https://xenitsupport.jira.com/browse/ALFREDAPI-428): Replace PersonService#getPerson(String userName) with getPersonOrNull(String userName) to avoid creation of new users when getPerson is called with a nonexistent userName.
 
+
 ### Deleted
 
 
-
 ## 2.5.2 (2020-08-12)
+
+### Fixed
+* [ALFREDAPI-338](https://xenitsupport.jira.com/browse/ALFREDAPI-338): Fixed issue when the provided name would not be set while copying a node
+
+## 2.5.1 (2020-08-05)
 
 ### Fixed
 * [ALFREDAPI-338](https://xenitsupport.jira.com/browse/ALFREDAPI-338): Fixed issue when the provided name would not be set while copying a node